CR separate hkubeAPI from streamingManager
