+++ hkube_python_wrapper/wrapper/algorunner.py
@@ -10,6 +10,7 @@ from .messages import messages
 from hkube_python_wrapper.tracing import Tracer
 from hkube_python_wrapper.codeApi.hkube_api import HKubeApi
 from hkube_python_wrapper.communication.DataServer import DataServer
+from hkube_python_wrapper.communication.streaming.StreamingManager import StreamingManager
 from hkube_python_wrapper.util.queueImpl import Queue, Empty
 from hkube_python_wrapper.util.timerImpl import Timer
 import os
@@ -30,6 +31,7 @@ class Algorunner(DaemonThread):
         self._loadAlgorithmError = None
         self._connected = False
         self._hkubeApi = None
+        self.streamingManager = None
         self._msg_queue = Queue()
         self._dataAdapter = None
         self._dataServer = None
@@ -183,7 +185,8 @@ class Algorunner(DaemonThread):
 
         self._wsc = WebsocketClient(self._msg_queue, encoding, self._url)
         self._initStorage(options)
-        self._hkubeApi = HKubeApi(self._wsc, self, self._dataAdapter, self._storage)
+        self.streamingManager = StreamingManager()
+        self._hkubeApi = HKubeApi(self._wsc, self, self._dataAdapter, self._storage, self.streamingManager)
         self._registerToWorkerEvents()
 
         print('connecting to {url}'.format(url=self._url))
@@ -284,7 +287,7 @@ class Algorunner(DaemonThread):
     def _discovery_update(self, discovery):
         print('Got discovery update' + str(discovery))
         messageListenerConfig = {'encoding': config.discovery['encoding']}
-        self._hkubeApi.setupStreamingListeners(
+        self.streamingManager.setupStreamingListeners(
             messageListenerConfig, discovery, self._nodeName)
 
     def _setupStreamingProducer(self, me):
@@ -298,12 +301,12 @@ class Algorunner(DaemonThread):
         producerConfig['messagesMemoryBuff'] = config.discovery['streaming']['messagesMemoryBuff']
         producerConfig['encoding'] = config.discovery['encoding']
         producerConfig['statisticsInterval'] = config.discovery['streaming']['statisticsInterval']
-        self._hkubeApi.setupStreamingProducer(
+        self.streamingManager.setupStreamingProducer(
             onStatistics, producerConfig, self._input['childs'], me)
 
     def _start(self, options):
-        self._hkubeApi.setParsedFlows(self._input.get('parsedFlow'), self._input.get('defaultFlow'))
-        if (self.isStreamingPipeLine()):
+        self.streamingManager.setParsedFlows(self._input.get('parsedFlow'), self._input.get('defaultFlow'))
+        if (self.isStreamingPipeLine() and self._input['childs']):
             self._setupStreamingProducer(self._input.get("nodeName"))
         # pylint: disable=unused-argument
         span = None
@@ -333,6 +336,7 @@ class Algorunner(DaemonThread):
         except Exception as e:
             traceback.print_exc()
             Tracer.instance.finish_span(span, e)
+            self.runningStartThread = None
             self.sendError(e)
 
     def _handle_response(self, algorithmData, jobId, taskId, nodeName, savePaths, span):
@@ -423,7 +427,7 @@ class Algorunner(DaemonThread):
             sys.exit(code)
 
         except Exception as e:
-            print('Got error during exit: ' + e)
+            print('Got error during exit: ' + str(e))
             # pylint: disable=protected-access
             os._exit(0)
